pyproject.toml
[tool.poetry]
version = "0.1.0"
name = "AWS file manager"
description = "Microservice to implement a file system manager based on AWS with real-world quality."
authors = ["Pavel <prohorenkopavel74@gmail.com>"]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Internet",
    "Topic :: Software Development",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
    "Typing :: Typed",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Framework :: FastAPI",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
]

license = "LICENSE.txt"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
alembic = "^1.11.1"
asyncpg = "^0.27.0"
fastapi = "^0.96.0"
pydantic = "^1.10.9"
python-decouple = "^3.8"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.15"
uvicorn = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
skip = ["migrations"]

[tool.black]
exclude = '''
/(
  | migrations
)/
'''
line_length = 120
